Extract isBabelPluginFactory helper
